+++ apis/elasticsearch_client.rb
